Refactoring api part I
