+++ api/flaskr/weather.py
@@ -9,10 +9,15 @@ def get_current_weather(lat, lon):
     '''
     Method for getting weather at a location
     '''
-    endpoint = "https://api.openweathermap.org/data/2.5/"
-    current_weather_endpoint = endpoint + "onecall?"
-    payload = {'lat': str(lat), 'lon': str(lon), 'exclude': "current,minutely,hourly", 'appid': OPEN_WEATHER_API_KEY, 'units':'imperial'}
-    response = requests.get(current_weather_endpoint, payload)
+    endpoint = "https://api.openweathermap.org/data/2.5/onecall"
+    payload = {
+        'lat': str(lat), 
+        'lon': str(lon), 
+        'exclude': "current,minutely,hourly", 
+        'appid': OPEN_WEATHER_API_KEY, 
+        'units':'imperial'
+    }
+    response = requests.get(endpoint, payload)
     current_weather_data = response.json()
     return current_weather_data
